Sarcastibot using Discord.Py Client
This file utilizes the discord.py API to gather a message that starts with '$sarcastic', and camel case every other letter. 
After doing so, it sends the message back to the server, and deletes the original message sent by the user. 

It uses the client side API instead of the bot side.